Fixes e2e test
