+++ internal/configs/config.go
@@ -7,7 +7,6 @@ package configs
 
 import (
 	"fmt"
-	"github.com/hashicorp/hcl/v2/hclsyntax"
 	"log"
 	"sort"
 
@@ -924,110 +923,6 @@ func (c *Config) TransformForTest(run *TestRun, file *TestFile, evalCtx *hcl.Eva
 	}, diags
 }
 
-// TODO not sure if this is the right place for this struct
-var _ hcl.Body = &testProviderBody{}
-
-// testProviderBody is a wrapper around hcl.Body that allows us to use prepared EvalContext for provider configuration.
-type testProviderBody struct {
-	originalBody hcl.Body
-	evalCtx      *hcl.EvalContext
-}
-
-func (c testProviderBody) Content(schema *hcl.BodySchema) (*hcl.BodyContent, hcl.Diagnostics) {
-	content, diags := c.originalBody.Content(schema)
-	if diags.HasErrors() {
-		return nil, diags
-	}
-	attrs := content.Attributes
-	for name, attr := range attrs {
-		vars := attr.Expr.Variables()
-		containsRunRef := false
-		for _, v := range vars {
-			if v.RootName() == "run" {
-				containsRunRef = true
-				break
-			}
-		}
-		if !containsRunRef {
-			continue
-		}
-
-		attr, valueDiags := c.getLiteralAttr(attr)
-		diags = append(diags, valueDiags...)
-		if diags.HasErrors() {
-			continue
-		}
-		attrs[name] = attr
-	}
-	return content, diags
-}
-
-func (c testProviderBody) PartialContent(schema *hcl.BodySchema) (*hcl.BodyContent, hcl.Body, hcl.Diagnostics) {
-	var diags hcl.Diagnostics
-
-	partialContent, remain, diags := c.originalBody.PartialContent(schema)
-
-	if diags.HasErrors() {
-		return nil, nil, diags
-	}
-
-	for name, attr := range partialContent.Attributes {
-		vars := attr.Expr.Variables()
-		containsRunRef := false
-		for _, v := range vars {
-			if v.RootName() == "run" {
-				containsRunRef = true
-				break
-			}
-		}
-		if !containsRunRef {
-			continue
-		}
-		attr, valueDiags := c.getLiteralAttr(attr)
-		diags = append(diags, valueDiags...)
-		if diags.HasErrors() {
-			continue
-		}
-		partialContent.Attributes[name] = attr
-	}
-
-	return partialContent, testProviderBody{originalBody: remain, evalCtx: c.evalCtx}, diags
-
-}
-
-func (c testProviderBody) JustAttributes() (hcl.Attributes, hcl.Diagnostics) {
-	attrs, diags := c.originalBody.JustAttributes()
-	for name, attr := range attrs {
-		attr, valueDiags := c.getLiteralAttr(attr)
-		diags = append(diags, valueDiags...)
-		if diags.HasErrors() {
-			continue
-		}
-		attrs[name] = attr
-	}
-	return attrs, diags
-}
-
-func (c testProviderBody) MissingItemRange() hcl.Range {
-	return c.originalBody.MissingItemRange()
-}
-
-func (c testProviderBody) getLiteralAttr(attr *hcl.Attribute) (*hcl.Attribute, hcl.Diagnostics) {
-	val, diags := attr.Expr.Value(c.evalCtx)
-	if diags.HasErrors() {
-		return nil, diags
-	}
-	return &hcl.Attribute{
-		Name: attr.Name,
-		Expr: &hclsyntax.LiteralValueExpr{
-			Val:      val,
-			SrcRange: attr.Range,
-		},
-		Range:     attr.Range,
-		NameRange: attr.NameRange,
-	}, nil
-}
-
 func (c *Config) getProviderConfigTransformForTest(evalCtx *hcl.EvalContext) configTransformFunc {
 	return func(run *TestRun, file *TestFile) (func(), hcl.Diagnostics) {
 		var diags hcl.Diagnostics
@@ -1090,7 +985,9 @@ func (c *Config) getProviderConfigTransformForTest(evalCtx *hcl.EvalContext) con
 				}
 
 				providerConfig := testProvider.Config
-				if ctxRunOutputExists {
+				// If we have test run block output and provider config exists we can use testProviderBody to wrap it.
+				// It can be used to evaluate run block expressions inside provider config with prepared evalCtx.
+				if ctxRunOutputExists && providerConfig != nil {
 					providerConfig = testProviderBody{originalBody: providerConfig, evalCtx: evalCtx}
 				}
 
@@ -1112,7 +1009,9 @@ func (c *Config) getProviderConfigTransformForTest(evalCtx *hcl.EvalContext) con
 			// Otherwise, let's copy over and overwrite all providers specified by
 			// the test file itself.
 			for key, provider := range file.Providers {
-				if ctxRunOutputExists {
+				// If we have test run block output and provider config exists we can use testProviderBody to wrap it.
+				// It can be used to evaluate run block expressions inside provider config with prepared evalCtx.
+				if ctxRunOutputExists && provider.Config != nil {
 					provider.Config = testProviderBody{originalBody: provider.Config, evalCtx: evalCtx}
 				}
 				next[key] = provider
